removed some duplicate imports, integrated new DB system(still left car input upon login, for basic display upon login test)
