+++ app/forms/registration_form.py
@@ -1,14 +1,9 @@
 from flask_wtf import FlaskForm
-from wtforms import PasswordField, SubmitField, EmailField
+from wtforms import PasswordField, SubmitField, EmailField, BooleanField
 from wtforms.validators import DataRequired, Length, EqualTo, Email
 from app.models.password_validator import CustomPasswordValidator
 
-
-from flask_wtf import FlaskForm
-from wtforms import PasswordField, SubmitField, BooleanField
-from wtforms.fields import EmailField
-from wtforms.validators import DataRequired, Email, Length, EqualTo
-from app.models.password_validator import CustomPasswordValidator
+# from wtforms.fields import EmailField
 
 
 class RegistrationForm(FlaskForm):
